pyproject.toml
[project]
name = "turbulence"
version = "0.1.0"
description = "Workflow simulation and testing framework"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Turbulence Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "jsonpath-ng>=1.6.0",
    "jinja2>=3.1.0",
    "rich>=13.0",
    "jsonschema>=4.20.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "grpcio>=1.60.0",
    "grpcio-reflection>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
    "types-protobuf>=4.24.0",
]

[project.scripts]
turbulence = "turbulence.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/turbulence"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
src = ["src"]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "N",    # pep8-naming
    "S",    # flake8-bandit (security)
    "ANN",  # flake8-annotations
    "D",    # pydocstyle
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D107",   # Missing docstring in __init__
    "B008",   # Function call in default argument (standard Typer pattern)
    "ANN401", # Dynamically typed expressions (Any) - needed for Pydantic validators
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # Allow assert in tests
    "ARG002", # Allow unused method arguments in tests (fixtures)
    "D103",   # Missing docstring in public function
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "yaml",
    "jsonpath_ng",
    "jsonpath_ng.exceptions",
    "jsonschema",
    "jsonschema.validators",
    "referencing",
    "referencing.jsonschema",
]
ignore_missing_imports = true
